Fix Hulu Video Player Dropshadow
Added a fix for Hulu's video player's shadow. This one was a bit tricky as the controls were extremely hard to locate as they were obscured behind several confusing layers. But the Hulu gradient shadow was the most egregious as it covered nearly 40% of the screen when scrolling through the timeline.